Created Function Files for ease of access
Split up the randomWalkSimulator to individual python files to make understanding and finding of functions easier. Changed all files to have .py so that their functions can be imported when needed.

Added redundancy to the mean Displacement checker in case of human error in choosing max step length.

Fixed the labels on the 3D plot.