New method for motion and image generation
For run and tumble motion along with Brownian motion, a new method has been developed that doesn't store the coordinates. If you would like methods to store the coordinates to look at you will need to use code from this commit.

As this is no longer needed in my code, it will be removed from this point onwards and the new code type will be from here onwards.